Improve estimatation of N effect by increasing sampling density along interpolated rarefaction curve.
This has a major payoff for some of the case-study datasets which yielded bizarre non-zero centered null models. These changes result in null model results centered on zero as expected. This higher density sampling will result in longer compute times for datasets with 1e5 individuals or more and therefore a more elegant approach to sampling is likely still needed.