Deploy suppressWarnings(lgamma) on rarefaction function.
suppressWarnings must be used here because lgamma generates two different warnings which are both in
response to being evaluated over negative integers. Although the test in the ifelse statement is supposed to catch negative values when ifelse is run all of the values for "no" are computed but then filled back in with 0's therefore the lgamma warnings are triggered although their effects are countered by the "yes" value which is appropriately 0. Although this approach is ugly it is quite a bit faster then looping throug each species individually which completely avoids the warnings.