Refactor spatial rarefaction code.
No change to output estimates. The calculation of cumulative richness is less elegent but easier to understand, the calculation is also slightly faster for a small community and speed gains may increase for larger datasets. The speed gain results because a matrix not converted to data.frame and unncessary copies of matrices were removed.