Group-scale permutation test across all levels of groups
The mean difference across all levels of the group variable is compared against permutations to test if the observed groups are more different in biodiversity than expected due to chance. This is in contrast to the previous approach of comparing all other levels to the reference level and computing just the difference between reference and not-reference. This may be a useful test but it is fundamentally different than the test that is carried out at the sample scale using the anova F-statistic. The function argument 'ref_group' is removed because it is not needed.