Compute D_bar in get_mob_stats and add to plots
D_bar was previously computed only at the gamma scale as a test statistic for permutation tests. This commit adds D_bar as a measure of effect size that is also computed at the alpha scale. The F-statistic which is still the test statistic at the alpha scale is no longer output on the graphics because it is not as easily interpreted as D_bar which is now added to the graphics.