Modified function rarefaction to fix bug
when computing out spatial plot-based accumulation (i.e., method='spat') as part of the permutation routine in get_delta_stats the permuted community would be so sparse that the matrix was reduced to a vector because this line of code removed any columns from the community matrix where the species never occured. As far as I can tell this code was orginally place her just to decrease memory consumption so adding it back in shouldn't be a big deal.