Do not round sampling effort
When computing the sample-based non-spatial rarefaction curve the effort was being rounded so that the binomial coefficients would not issue warnings. The gamma formulation of the function allows non-integer efforts and these are now used. We do not expect this will change the shape of the curves much but these are technically more exact.