Major refactoring and extension of delta analysis
Both type discrete and continuous are working using the same code the only difference is in how the grouping variable is treated (factor or numeric respectively). get_delta_stats returns several statistics including y-intercept, slope, r2, and adj r2.

The refactoring relies heavily on dplyr, purrr, and tidyr. As dplyr is already a dependancy of the package this do not add a big install burden to users. Also Hadley Wickham assurred me that purrr::map is stable.