Improve the handling of the explanatory and group variables in the multiscale delta analysis.
Previously the argument "group_var" specified the explanatory varible and the grouping variable for the analysis. Now the argument "env_var" specifies the explanatory variable and "group_var" just specifies the grouping variable. This was necessary because sometimes two different groups have the exact same value of the explanatory variable and it is desireable to not conflate them. This solves issue #226