Corrected bug in sphere_dist and changed spatial dependancies.
Great circle distances did not appear to be correct. These are now computed by a corrected formula copied from fields::rdist.earth. The test associated with sphere_dist was removed as it is no longer needed. The dependancies on the packages sp and rgeos were replaced by geosphere which provides functions for computing arc distances and centroids that is more efficient and correct even for some edge cases but not all (polar regions).