Fix bug in rarefaction for nsSBR by jimjunker1
The bug occurs when the effort or number of indivdiauls is larger than observed in the group. In this case rarefaction should return either observed S or extrapolated S which is what this patch allows. The patch was sumbited via: https://github.com/MoBiodiv/mobr/pull/265 and the bug was documented: https://github.com/MoBiodiv/mobr/issues/264